requirements.txt
pokemontcgsdk==3.4.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
python-telegram-bot==20.0a0
SQLAlchemy==1.4.36